reformatted project structure
