+++ const.go
@@ -1,0 +1,7 @@
+package gosupreme
+
+// Endpoint
+const (
+	StockURL = "https://supremenewyork.com/mobile_stock.json"
+	StyleURL = "https://supremenewyork.com/shop/%d.json"
+)
