+++ go.mod
@@ -1,8 +1,3 @@
 module github.com/doxify/supreme-go-api
 
 go 1.16
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/kr/pretty v0.2.1 // indirect
-)
